JOB PORTAl
This is a simple job portal where a jobseeker can register and
get to know about new jobs.

Jobseeker has to create an account before he can apply for the
jobs he is interested in.

In case a jobseeker  forgets his password he can click on the
forget password option and the password will be send to his mail
account.